test: move all relevant tests to the transform test file
