--- test/snapshots/transform-feature.test.js.snapshot
@@ -1,11 +1,0 @@
-exports[`should transform TypeScript class fields 1`] = `
-"class Counter {\\n    count = 0;\\n    increment() {\\n        this.count++;\\n    }\\n}\\n"
-`;
-
-exports[`should transform TypeScript private class fields 1`] = `
-"class Counter {\\n    #count = 0;\\n    increment() {\\n        this.#count++;\\n    }\\n    getCount() {\\n        return this.#count;\\n    }\\n}\\n"
-`;
-
-exports[`should transform TypeScript type annotations and type guards 1`] = `
-"function isString(value) {\\n    return typeof value === 'string';\\n}\\n"
-`;
